Adjust layer initialization
The initialization method implemented in `MACLayer.reset_parameters` is
based on He et al. (2015), but was adapted to take the hardware gain
into account and uses a truncated normal distribution.
If not specified, the number of sends is adjusted such that at least 3
standard deviations of the initial weights can be represented, which is
a reasonable default in most cases.

Change-Id: Id0404b18249652fb230efcbc0f9b31b3c21e26ba