Fix infrequently failing expanded conv test
The expansion of convolution kernels can cause an overfull synapse
matrix that gets split up into multiple MAC operations. Each of those
will potentially contribute with a rounding error up to 1. The
previously used relative tolerance was mostly sufficient, but not
always. Now, we calculate how many synapse matrices are used during the
operation, which is the maximum expected rounding error, and allow this
as an absolute tolerance for the result.

We also now generate more test cases for the conv1d operation.

Change-Id: Icbce2d495870a0097a71743d4ac7a8611a78c943